requirements.txt
Django==2.2.14
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
#mysqlclient==1.4.6
coreapi==2.3.3
django-filter==2.3.0
xlrd==1.2.0
drf-yasg==1.17.1
